+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ChangeLog
 
 ## [Unreleased]
+### Add
+- start to use database migration
 
 ## [1.0.1] 2018-12-24
 ### Chanage